+++ packages/client/src/app/pages/index.ts
@@ -2,3 +2,4 @@ export { BrowseComponent } from './browse/browse.component';
 export { GroupsComponent } from './groups/groups.component';
 export { NewsComponent } from './news/news.component';
 export { DocsPageComponent } from './docs-page/docs-page.component';
+export { PostPageComponent } from './post-page/post-page.component';
