--- packages/client/src/app/pages/news/index.ts
@@ -1,2 +1,0 @@
-export { NewsComponent } from './news.component';
-export { PostPageComponent } from './post-page/post-page.component';
