+++ packages/client/src/app/pages/post-page/post-page.component.ts
